chore: organize TransactionController wallet init files
