+++ app/scripts/lib/transaction/containers/util.test.ts
@@ -10,7 +10,7 @@ import {
   Messenger,
   MockAnyNamespace,
 } from '@metamask/messenger';
-import { TransactionControllerInitMessenger } from '../../../wallet-init/instance-options/transaction-controller-messenger';
+import { TransactionControllerInitMessenger } from '../../../wallet-init/messengers/transaction-controller-messenger';
 import { applyTransactionContainers } from './util';
 import { enforceSimulations } from './enforced-simulations';
