+++ app/scripts/lib/transaction/containers/util.ts
@@ -5,7 +5,7 @@ import {
 import { cloneDeep } from 'lodash';
 import { createProjectLogger } from '@metamask/utils';
 import type { Hex } from 'viem';
-import { TransactionControllerInitMessenger } from '../../../wallet-init/instance-options/transaction-controller-messenger';
+import { TransactionControllerInitMessenger } from '../../../wallet-init/messengers/transaction-controller-messenger';
 import { enforceSimulations } from './enforced-simulations';
 
 const log = createProjectLogger('transaction-containers');
