+++ app/scripts/lib/transaction/hooks/delegation-7702-publish.ts
@@ -13,7 +13,7 @@ import {
   findAtomicBatchSupportForChain,
   checkEip7702Support,
 } from '../../../../../shared/lib/eip7702-support-utils';
-import { TransactionControllerInitMessenger } from '../../../wallet-init/instance-options/transaction-controller-messenger';
+import { TransactionControllerInitMessenger } from '../../../wallet-init/messengers/transaction-controller-messenger';
 import {
   RelayStatus,
   RelaySubmitRequest,
